made controller class
